backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.4
pydantic-settings==2.7.0
firebase-admin==6.6.0
google-cloud-firestore==2.19.0
google-cloud-secret-manager==2.21.1
google-cloud-aiplatform==1.75.0
google-generativeai>=0.8.0
google-genai>=1.0.0
httpx==0.27.2
yfinance==0.2.51
mftool==2.8
python-multipart==0.0.20
PyPDF2>=3.0.0
pytest==8.3.4
